requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile dev.in
#
blinker==1.6.2
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   flask
build==0.10.0
    # via pip-tools
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   flask
    #   pip-tools
coverage==7.2.5
    # via -r dev.in
distlib==0.3.6
    # via virtualenv
filelock==3.12.0
    # via virtualenv
flake8==6.0.0
    # via -r dev.in
flask==2.3.2
    # via
    #   -c requirements.txt
    #   -r requirements.txt
identify==2.5.24
    # via pre-commit
itsdangerous==2.1.2
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   flask
jinja2==3.1.2
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   flask
markupsafe==2.1.2
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
nodeenv==1.7.0
    # via pre-commit
packaging==23.1
    # via build
pip-tools==6.13.0
    # via -r dev.in
platformdirs==3.5.0
    # via virtualenv
pre-commit==3.3.1
    # via -r dev.in
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyproject-hooks==1.0.0
    # via build
pyyaml==6.0
    # via pre-commit
tomli==2.0.1
    # via build
virtualenv==20.23.0
    # via pre-commit
waitress==2.1.2
    # via
    #   -c requirements.txt
    #   -r requirements.txt
werkzeug==2.3.3
    # via
    #   -c requirements.txt
    #   -r requirements.txt
    #   flask
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
